+++ src/test/preferences.test.ts
@@ -4,7 +4,7 @@ import * as kotlinExt from '../extension';
 // import * as compliance from '../compliance';
 // import * as commands from '../commands';
 // import * as file_generator from '../file_generator';
-import * as preferences from '../preferences';
+import * as preferences from '../util/preferences';
 // import * as template_interpreter from '../template_interpreter';
 // import * as path from "path";
 import * as fs from 'fs';
