+++ src/test/template_interpreter.test.ts
@@ -5,7 +5,7 @@ import * as assert from 'assert';
 // import * as commands from '../commands';
 // import * as file_generator from '../file_generator';
 // import * as preferences from '../preferences';
-import * as template_interpreter from '../template_interpreter';
+import * as template_interpreter from '../templates/template_interpreter';
 // import * as path from "path";
 // import * as fs from 'fs';
 import * as testingConsts from "./testingConstants";
