+++ src/util/compliance.ts
@@ -1,11 +1,11 @@
 "use strict";
 import * as vscode from "vscode";
-import { targetGradleRioVersion } from "./constants";
-import { createMainKt } from "./commands";
-import { createFileWithContent } from "./file_generator";
+import { targetGradleRioVersion } from "../constants";
+import { createMainKt } from "../commands";
+import { createFileWithContent } from "../file_manipulation/file_generator";
 import * as preferences from "./preferences";
 import * as fs from "fs";
-import * as kotlinExt from "./extension";
+import * as kotlinExt from "../extension";
 
 export function isGradleRioVersionCompliant(): boolean {
     console.log("Checking build.gradle compliance");
